- 좋아요 갯수, 댓글 갯수 추가

+++ src/main/java/com/move/review/repository/CommentRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface CommentRepository extends JpaRepository<Comment, Long> {
   List<Comment> findAllByReview(Review review);
-  int countByReview(Review review);
+  Long countByReview(Review review);
 }
