+++ src/main/java/com/move/review/repository/HeartRepository.java
@@ -9,5 +9,5 @@ import java.util.Optional;
 
 public interface HeartRepository extends JpaRepository<Heart, Long> {
     Optional<Heart> findByMemberAndReview(Member member, Review review);
-    int countByReview(Review review);
+    Long countByReview(Review review);
 }
